feat(publishers): add auth gates and deletion of all books of publisher on publisher deletion

+++ backend/app/Http/Controllers/PublisherController.php
@@ -6,11 +6,15 @@ use App\Http\Requests\StorePublisherRequest;
 use App\Http\Requests\UpdatePublisherRequest;
 use App\Http\Resources\PublisherResource;
 use App\Models\Publisher;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Response;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class PublisherController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -25,6 +29,7 @@ class PublisherController extends Controller
      */
     public function store(StorePublisherRequest $request): JsonResource
     {
+        $this->authorize('manager', $request->user());
         $publisher = Publisher::create($request->validated())->load(['books.authors', 'books.genre']);
         return new PublisherResource($publisher);
     }
@@ -42,6 +47,7 @@ class PublisherController extends Controller
      */
     public function update(UpdatePublisherRequest $request, Publisher $publisher): JsonResource
     {
+        $this->authorize('manager', $request->user());
         $publisher->update($request->validated());
         return new PublisherResource($publisher->load(['books.authors', 'books.genre']));
     }
@@ -49,8 +55,10 @@ class PublisherController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Publisher $publisher): Response
+    public function destroy(Publisher $publisher, Request $request): Response
     {
+        $this->authorize('manager', $request->user());
+        $publisher->books()->delete();
         return $publisher->delete() ? response()->noContent() : abort(500);
     }
 }
